==> new_demography/2.5 Cost calcs.Rmd <==
---
title: "NCD_countdown_costing"
author: "Yoshito Kawakatsu"
date: "2020年4月13日"
output: html_document
---

# Install the package
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls()) #Remove all
library(dplyr)
library(ggplot2)
library(readxl)
library(tidyverse)
library(kableExtra)
library(reshape2)
library(stringr)
library(zoo)

'%!in%' <- function(x,y)!('%in%'(x,y))
```

# Functions
```{r}
tradable_conversion <- function(unit_cost, tradable_ratio, currency, global_inflation, exchange){
  if(currency == "USD") {
    unit_cost*tradable_ratio*global_inflation} else {
      ((unit_cost*tradable_ratio)/exchange)*global_inflation
    }
}

nontradable_conversion <- function(unit_cost,tradable_ratio, currency, cpi_adjust, exchange,country, exchange_end, gni, gni_selected_country){
  if(currency == "USD" & (country == "LIC" | country == "LMIC" | country == "LIC+LMIC+UMIC")) {
    unit_cost*(1-tradable_ratio)*cpi_adjust*(gni_selected_country/gni)
  } else if (currency == "USD" & (country != "LIC" & country != "LMIC" & country != "LIC+LMIC+UMIC")){
    ((unit_cost*exchange)*(1-tradable_ratio))*cpi_adjust/exchange_end*(gni_selected_country/gni)} 
  else {
      ((unit_cost*(1-tradable_ratio)*cpi_adjust)/exchange_end)*(gni_selected_country/gni)
    }
}
```

# Install the necessary input files
```{r}

# population in need (Output from Will)
load("../new_demography/output/icer_rank_output_0830.Rda")

coverage<-read.csv("../new_inputs/Coverage0621.csv", stringsAsFactors = F)

# Unit cost adjustment
uc <- read.csv("../new_inputs/PINandCosts0625.csv", stringsAsFactors=F)
uc$unique_id <- paste0("C",uc$NCD,"_",uc$sub_id)

#table(uc$unique_id)
uc <- uc %>%filter(Original.Currency != "NA")

# Necessary files
# Exchange rate
exchange <- read_xlsx("../new_inputs/Exchange_rate.xlsx", na = "..")
exchange<-exchange[,c(1:60)]
ex2<-read_xlsx("../new_inputs/Exchange_rate_2020.xlsx", skip=3)
ex2<-ex2[,c(1,61,62)]

exchange<-left_join(exchange, ex2, by="Country")
exchange <- exchange%>%gather(year, value, -Country)%>%group_by(Country)%>%fill(value)%>%spread(year, value)


# Regional inflation
#regional_inflation <- read_xls("../new_inputs/Regional_inflation_20200212.xls", sheet = "Data")
regional_inflation <- read_xls("../new_inputs/Regional_inflation_2020.xls", sheet = "Data", skip=3)
#regional_inflation[,c(5:64)] <- (regional_inflation[,c(5:64)]/100)+1
regional_inflation[,c(5:65)] <- (regional_inflation[,c(5:65)]/100)+1

# COUntry grouping
class <- read.csv("../new_inputs/Country_groupings.csv")

#Country list: only eligible countries
table(class$NCD_region)
country_list <- class %>% 
  filter(World_bank == "LIC" | World_bank == "LMIC"| World_bank == "UMIC") %>% 
  filter(NCD_region %in% c("Latin America and Caribbean",
         "Central and Eastern Europe",
         "Central Asia, Middle East and North Africa",
         "Sub-Saharan Africa",
         "South Asia",
         "East and South East Asia",
         "Oceania")) %>% 
  select(World_bank, NCD_region ,Country, iso3)


# CPI adjusted
cpi <- read.csv("../new_inputs/CPI_WB_2018_added2.csv", stringsAsFactors=F)
cpi<-cpi[,-c(66:71)]
#add 2020 data
cpi2<-read.csv("../new_inputs/CPI_2020.csv", stringsAsFactors=F)
names(cpi2)[1]<-"Country"
cpi2<-cpi2[,c(1,65)]
cpi<-left_join(cpi, cpi2, by="Country")

cpi <- cpi[,c(3,6:66)]%>%gather(year, value, -Country)%>%group_by(Country)%>%fill(value)%>%spread(year, value)

#also add France
#france<-read.csv("../new_inputs/CPI_2020.csv", stringsAsFactors=F)%>%filter(Country.Code=="FRA")
#names(france)[1]<-"Country"
#cpi<-bind_rows(cpi, france[,-c(2,3,4)])

# GNI
GNI <- read.csv("../new_inputs/GNI_WB_USD_2017_re2.csv", stringsAsFactors=F)
GNI<-GNI[,-c(1,4,5,6,66,67,68)]
GNI2 <- read.csv("../new_inputs/GNI_2020.csv", stringsAsFactors=F)
GNI2<-GNI2[,c(1,64,65)]
names(GNI2)[1]<-"Country"

GNI<-left_join(GNI, GNI2, by="Country")
GNI <- GNI%>%gather(year, value, -Country, -Country.Code)%>%group_by(Country.Code)%>%fill(value)%>%spread(year, value)

```

# Data preprocessing
```{r}
head(all.pin)
head(all.dalys)

all.pin$unique_id <- paste0(all.pin$Code,"_",all.pin$sub_id)

#write.csv(all.pin, "NCD_data_0706/test_output/pin_all_icer.csv")
#write.csv(all.dalys,"NCD_data_0706/test_output/NCD_benefits_icer.csv")

all.pin.icer <- all.pin

#country list of no alcohol
noalc.list<-c("Afghanistan", "Bangladesh", "Iran", "Libya", "Sudan", "Yemen")

#list of alcohol policy unique_ids
intersectoral.alc<-c("5.2_a", "5.4_a")

###########################################  
##update PIN for intersectoral policies##
############################################

#convert intersectoral PIN back to total pop for years 2023-2030
all.pin.icer$pin[all.pin.icer$group=="Adjsuted" & all.pin.icer$Code>5 & all.pin.icer$year_id>=2023]<- all.pin.icer$pin[all.pin.icer$group=="Adjsuted" & all.pin.icer$Code>5& all.pin.icer$year_id>=2023]/all.pin.icer$Coverage[all.pin.icer$group=="Adjsuted" & all.pin.icer$Code>5& all.pin.icer$year_id>=2023]

#scale-up coverage (1/3 in 2023, 2/3 in 2024, full coverage 2025-2030)
all.pin.icer$pin[all.pin.icer$group=="Adjsuted" & all.pin.icer$Code>5 & all.pin.icer$year_id==2023]<- (1-(all.pin.icer$Coverage[all.pin.icer$group=="Adjusted" & all.pin.icer$Code>5 & all.pin.icer$year_id==2023]-0.1))*(1/3)


##edit for countries in which alcohol is banned##
all.pin.icer$pin[all.pin.icer$unique_id%in%intersectoral.alc & all.pin.icer$location_name%in%noalc.list]<-0

benefits <- merge(y = all.dalys, x = class, by.x = "Country", by.y = "location_name")

changeq30_old<-all.q30%>%filter(year_id==2023 | year_id==2030)%>%
  select(c(year_id, Adjusted, location_name, Code))%>%
  spread(year_id, Adjusted)%>%mutate(q30diff=((`2023`-`2030`)/`2023`)*100)%>%
  rename(Country = location_name)%>%select(c(Country, Code, q30diff))

changeq30<-all.q30%>%group_by(Code, location_name)%>%summarise(q30diff=100*sum(q30.ave/Baseline))%>%
  rename(Country = location_name)%>%select(c(Country, Code, q30diff))
```

# Settings
```{r}
#Setting
#tradable_ratio <- 0.3
UC_Years <- 2020  # The output year
# ancillary_HF_cost <- 0.5 # possible change
# above_HF <- 0.17 # possible change
```

# Data frame
```{r}
final_all <- data.frame(Code = character() ,
                        adjusted_uc = numeric(),
                        pin_sum= numeric(),
                        year_id =  numeric(),
                        total_cost = numeric(),
                        senario = character() ,
                        Country = character(),
                        Class = character(),
                        Region = character(),
                        stringsAsFactors=FALSE)


final_all_intervention <- data.frame(Code = character(),
                                     Intervention = character(),
                                     senario = character() ,
                                     Country = character(),
                                     adjusted_uc = numeric(),
                                     pin_sum = numeric(),
                                     total_cost = numeric(),
                                     Region = character(),
                                     stringsAsFactors=FALSE) 




final_all_ICER <- data.frame(Code = character(),
                             Intervention = character(),
                             Country = character(),
                             adjusted_uc = numeric(),
                             pin_sum = numeric(),
                             total_cost = numeric(),
                             pin_sum_adjusted = numeric(),
                             total_cost_adjusted = numeric(),
                             increment = numeric(),
                             Super_region = character(),
                             NCD_region = character(),
                             SDI = character(),
                             World_bank = character(),
                             iso3 = character(),
                             period = character(),
                             life_years_gained = numeric(),
                             deaths_averted = numeric(),
                             q30diff=numeric(),
                             ICER = numeric(),
                             ICER2 = numeric(),
                             ICER_rank = numeric(),
                             ICER_rank2 = numeric()
)

increment_all_year      <- data.frame(
                             unique_id = character(),
                             year_id = numeric(),
                             Class = character(),
                             Region = character(),
                             adjusted_uc = numeric(),
                             base_pin = numeric(),
                             base_total = numeric(),
                             adjust_pin = numeric(),
                             adjust_total = numeric(),
                             increment = numeric()
)

```

# Calculate adjusted costs & total by interventions
```{r}
i <- 1
c <- 1
n <- 1
y <- 1

for (c in c(1:85, 87:124)){

  selected_country <- as.character(country_list$Country[c])
  selected_country_code <- as.character(country_list$iso3[c])
  selected_country_category <- as.character(country_list$World_bank[c])
  selected_country_region <- as.character(country_list$NCD_region[c])

#Adjusted UC start------------------------------------------
# Adjustment database
    no <- nrow(uc)
    
    adjustment <- data.frame(ID = 1:no, 
                             country = NA, 
                             year = NA, 
                             n = NA, 
                             cpi_study = NA, cpi_end = NA, cpi_adjust = NA,
                             global_inflation = NA,
                             exchange = NA,
                             exchange_end = NA, 
                             #gni_selected_country = NA,
                             gni = NA)

    
    for (i in 1:no){
      
      #Country
      country <- as.character(uc[i,"Country"])
      adjustment[i,"country"] <- country
      
      #Original UC_Year
      year <- uc[i,"Year"]
      adjustment[i,"year"] <- year
      
      # Exchange of original UC
      n <- as.numeric(year) - 1960 +2
      end_year <- as.numeric(UC_Years - 1960 +2)
      
      adjustment[i,"n"] <- n
      adjustment[i,"exchange"] <- as.numeric(exchange[exchange$Country == country, n])
      
      # Exchange rate in 2017
      adjustment[i,"exchange_end"] <- as.numeric(exchange[exchange$Country == country, end_year])
      
      # consumer price index
      study_year <- as.numeric(year) - 1960 + 2
      end_year <- as.numeric(UC_Years - 1960 + 2)

      adjustment[i,"cpi_study"] <- as.numeric(cpi[cpi$Country == country, study_year])
      adjustment[i,"cpi_end"] <- as.numeric(cpi[cpi$Country == country, end_year])
      
      # Global inflation calculation
      col_n <- as.numeric(year) - 1960+5
      col_end <- as.numeric(UC_Years) - 1960+5
      
      adjustment[i,"global_inflation"] <-
        as.numeric(prod(as.matrix(regional_inflation[regional_inflation$Country == "World", col_n:col_end])))
      
      #GNI
      GNI_year <- UC_Years - 1960 +3
      adjustment[i,"gni"] <- as.numeric(GNI[GNI$Country == country, GNI_year])
      
      # adjustment[i,"gni"] <- ifelse(selected_country_category == "UMIC", 
      #                               as.numeric(GNI[GNI$Country == "UMIC", GNI_year]),
      #                               adjustment[i,"gni"])
      
      
    }
    
      # adjustment$gni <- ifelse(adjustment$country == "LMIC", 
      #                            as.numeric(GNI[GNI$Country == "UMIC", GNI_year]),
      #                              adjustment$gni)
    
    # GNI adjustment
    adjustment$gni_selected_country <- as.numeric(GNI[GNI$Country.Code == selected_country_code, GNI_year])
    adjustment$gni_adjust <- adjustment$gni_selected_country/ adjustment$gni
    
    # CPI adjustment rate
    adjustment[,"cpi_adjust"] <- adjustment$cpi_end / adjustment$cpi_study
    
    # Merge adjustments with uc
    ucm <- cbind(uc, adjustment)
    
    # Tradable
    ucm$tradable_ratio <- ucm$Traded

    ucm$tradable_uc <- NA
    for (i in 1:no){

        ucm$tradable_uc[i] <- tradable_conversion(
                               unit_cost = as.numeric(ucm$Original.Unit.Cost[i]),
                               tradable_ratio = ucm$tradable_ratio[i],
                               currency = ucm$Original.Currency[i],
                               global_inflation = ucm$global_inflation[i],
                               exchange = as.numeric(ucm$exchange[i]))
                    }
    
    ucm$nontradable_uc <- NA    
     for (i in 1:no){
      ucm$nontradable_uc[i] <- nontradable_conversion(
                               unit_cost = as.numeric(ucm$Original.Unit.Cost[i]),
                               tradable_ratio = ucm$tradable_ratio[i],
                               currency = ucm$Original.Currency[i],
                               cpi_adjust = ucm$cpi_adjust[i],
                               country = ucm$country[i],
                               exchange = ucm$exchange[i],
                               exchange_end = ucm$exchange_end[i],
                               gni = ucm$gni[i],
                               gni_selected_country = ucm$gni_selected_country[i])
                    }
    
    # Add ancillary + above HF costs

     ucm$adjusted_uc <- (ucm$tradable_uc + ucm$nontradable_uc)*(ucm$Cost.adjustment)

     # Class check
     country_class <- as.character(country_list[country_list$Country == selected_country,]$World_bank)
     
     adjusted_uc <- if(country_class == "LIC"){
              ucm %>% 
                filter(unique_id != "C4.1_a")
     } else {
              ucm
     }

     #subtract adjusted CVD treatment costs from diabetes treatment cost (apologies, this is hideous code)
     adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="c"]<-adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="c"]- adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.4 & adjusted_uc$sub_id=="e"]
     
     adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="d"]<-adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="d"]- adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.4 & adjusted_uc$sub_id=="e"]
     
     adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="e"]<-adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="e"]- adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.4 & adjusted_uc$sub_id=="e"]
     
     adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="f"]<-adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.3 & adjusted_uc$sub_id=="f"]- adjusted_uc$adjusted_uc[adjusted_uc$NCD==2.4 & adjusted_uc$sub_id=="e"]
     
     # Output: adjusted unit cost
     file.name <- paste0("../new_demography/output/costs/bycountry/",selected_country,"_adjusted_uc_",UC_Years,".csv")
     write.csv(adjusted_uc, file.name) 

     # Merge interventions costs
     adjusted_uc_sum <- adjusted_uc %>% group_by(NCD) %>% 
       summarize(adjusted_uc = sum(adjusted_uc))
     
     # Output: adjusted & merged unit cost
     #file.name <- paste0("../new_demography/output/costs/",selected_country,"_adjusted_uc_merged_",UC_Years,".csv")
     #write.csv(adjusted_uc_sum, file.name)     
        
# Adjusted UC end ------------------------------------------
     
    result.df <- adjusted_uc %>% select(unique_id)
    result.df2 <- adjusted_uc %>% select(unique_id)

    
# full data -----------------------
  country.pop <- all.pin.icer %>% 
  filter(location_name == selected_country)
    
  full.data <- merge.data.frame(x = country.pop, y = adjusted_uc[,c("unique_id", "Original.Unit.Cost", "Cost.adjustment", "Traded", "tradable_uc", "nontradable_uc", "adjusted_uc" )], by = "unique_id", all.x = T)

  full.data$total_cost <- full.data$pin * full.data$adjusted_uc
  
  file.name <- paste0("../new_demography/output/costs/",selected_country,"_all_by_intervention_year",".csv")
  #write.csv(full.data, file.name)
  
  
adjusted_uc$unique_id<-sub('.','',adjusted_uc$unique_id)
result.df$unique_id<-sub('.','',result.df$unique_id)
result.df2$unique_id<-sub('.','',result.df2$unique_id)

# Baseline population in need------------------
for (y in 3:10){
##PIN by intervention
# pin_group = all code
pin_group <- all.pin.icer %>% 
  filter(location_name == selected_country,
         year_id == 2020 + y, 
         group == "Baseline") %>% 
  dplyr::group_by(unique_id) %>% dplyr::summarize(pin_sum = sum(pin))

pin_group$year_id <- 2020 + y

uc_pin <- merge.data.frame(adjusted_uc, pin_group, by = "unique_id")

uc_pin <- uc_pin %>% 
  dplyr::mutate(total_cost = adjusted_uc * pin_sum)

uc_pin_all <- if(y == 3) {
  uc_pin
} else {
  rbind(uc_pin_all, uc_pin)
}

result.df <- merge(result.df, uc_pin[,c("unique_id","total_cost")], by = "unique_id")

label <- paste0("cost_",2020+y)
names(result.df)[ which( names(result.df)=="total_cost" ) ] <- label
}

  
# By Interventions
final.df1_int <- result.df
final.df1_int$senario <- "Baseline"
final.df1_int$Country <- selected_country
final.df1_int$Class <- country_class
final.df1_int$Region <- selected_country_region

final.df1_int$total_cost <- apply(final.df1_int[,2:9],1,sum)

# Add unit cost
final.df1_int <- merge(final.df1_int, uc_pin[,c("unique_id", "adjusted_uc")] , by = "unique_id")

# Add pin
pin_code <- all.pin.icer %>% 
  filter(location_name == selected_country,
         group == "Baseline") %>% 
  group_by(unique_id) %>% 
  summarize(pin_sum = sum(pin))

final.df1_int <- merge(final.df1_int, pin_code[,c("unique_id", "pin_sum")], by = "unique_id")

 # final.df1_int <- final.df1_int %>%
 #   select(unique_id, senario, Country, adjusted_uc, pin_sum ,total_cost, Region)
 # 

# years
result.df1_long <- uc_pin_all

result.df1_long$senario <- "Baseline"
result.df1_long$Country <- selected_country
result.df1_long$Class <- country_class
result.df1_long$Region <- selected_country_region


#Adjustment-------------

for (y in 3:10){

# pin_group = all code
pin_group <- all.pin.icer %>% 
  filter(location_name == selected_country,
         year_id == 2020 + y, 
         group == "Adjusted") %>% 
  dplyr::group_by(unique_id) %>% 
  dplyr::summarize(pin_sum = sum(pin))

pin_group$year_id <- 2020 + y

uc_pin <- merge.data.frame(adjusted_uc, pin_group, by = "unique_id")


uc_pin <- uc_pin %>% 
  mutate(total_cost = adjusted_uc * pin_sum)

uc_pin_all <- if(y == 3) {
  uc_pin
} else {
  rbind(uc_pin_all, uc_pin)
}

result.df2 <- merge(result.df2, uc_pin[,c("unique_id","total_cost")], by = "unique_id")

label <- paste0("cost_",2020+y)
names(result.df2)[ which( names(result.df2)=="total_cost" ) ] <- label

}

# By intervention & save

final.df2_int <- result.df2
final.df2_int$senario <- "Adjusted"
final.df2_int$Country <- selected_country
final.df2_int$Class <- country_class
final.df2_int$Region <- selected_country_region

final.df2_int$total_cost <- apply(final.df2_int[,2:9],1,sum)

# Add unit cost
final.df2_int <- merge(final.df2_int, uc_pin[,c("unique_id", "adjusted_uc")])

# Add pin  #why this?
pin_code <- all.pin.icer %>% 
  filter(location_name == selected_country, 
         group == "Adjusted") %>% 
  group_by(unique_id) %>% 
  summarize(pin_sum = sum(pin))

final.df2_int <- merge(final.df2_int, pin_code[,c("unique_id", "pin_sum")])

# 
#  final.df2_int <- final.df2_int %>% 
#    select(Code, senario, Country, adjusted_uc, pin_sum ,total_cost, Region)
  
final_country_tc_int <- rbind(final.df1_int,final.df2_int)


#Final dataset by interventions-------------------
file.name <- paste0("../new_demography/output/costs/",selected_country,"_total_cost_2030_intervention",".csv")
#write.csv(final_country_tc_int, file.name)

final_all_intervention <- rbind(final_all_intervention, final_country_tc_int)
#--------------------------------------------------

# By year
result.df2_long <- uc_pin_all

result.df2_long$senario <- "Adjusted"
result.df2_long$Country <- selected_country
result.df2_long$Class <- country_class
result.df2_long$Region <- selected_country_region

# Save
final_country_tc3 <- rbind(result.df1_long,result.df2_long)
file.name <- paste0("../new_demography/output/costs/",selected_country,"_total_cost_2030",".csv")
#write.csv(final_country_tc3, file.name)

#Final dataset-------------------
  final_all <- rbind(final_all, final_country_tc3)

# ----------------------
# incremental cost by year and intervention
base_result <- result.df1_long %>% 
  filter(senario == "Baseline")%>% 
  select(unique_id, Country, Class, Region, year_id, adjusted_uc, pin_sum, total_cost) %>% 
  rename(base_pin = pin_sum,
         base_total = total_cost) 

adjust_result <- result.df2_long %>%
  filter(senario == "Adjusted") %>% 
  select(unique_id, year_id, pin_sum, total_cost) %>% 
  rename(adjust_pin = pin_sum,
         adjust_total = total_cost) 


increment_by_year <- merge(base_result, adjust_result , by = c("unique_id", "year_id"))

increment_by_year$incremental <- increment_by_year$adjust_total - increment_by_year$base_total

file.name <- paste0("../new_demography/output/costs/",selected_country,"_increment_cost_2023_2030",".csv")
#write.csv(increment_by_year, file.name)

increment_all_year <- rbind(increment_all_year, increment_by_year)

# ICER calculation
#  final.df2_int2 <- final.df2_int %>% 
#    rename(pin_sum_adjusted = pin_sum,
#           total_cost_adjusted = total_cost) %>% 
#    select(unique_id, senario, pin_sum_adjusted, total_cost_adjusted)
#  
# country.all <- merge(final.df1_int,final.df2_int2, by = "unique_id")
# country.all$increment <- country.all$total_cost_adjusted - country.all$total_cost


country.all.id <- colsplit(increment_by_year$unique_id,"_",names=c("Code","sub_id"))
# From unique_id to Code
increment_by_year$Code <- country.all.id$Code
increment_by_year$sub_id <- country.all.id$sub_id


country.all.sum <- increment_by_year %>% 
  group_by(Code,year_id) %>% 
  summarize(sum_increment = sum(incremental))


# Benefit
benefits.country <- benefits %>% 
  filter(Country == selected_country) %>% 
  select(-Region, -Country) %>% 
  group_by(Code, year_id) %>% 
  summarize(sum_ly = sum(DALY.ave))

country.all2 <- merge(x = country.all.sum, y = benefits.country, by = c("Code", "year_id"),all.x = T)

country.all2$Country <- selected_country

country.all2 <- country.all2 %>% 
  mutate(ICER = sum_increment / sum_ly
         )

file.name <- paste0("../new_demography/output/costs/",selected_country,"_ICER_2023_2030",".csv")
#write.csv(country.all2, file.name)


country.all3 <- country.all2 %>% 
  group_by(Code) %>% 
  summarise(sum_increment_all = sum(sum_increment),
            sum_ly_all = sum(sum_ly)) %>% 
  mutate(
    ICER = sum_increment_all / sum_ly_all,
    ICER_rank = rank(ICER)
  ) %>% 
  mutate(
    Country = selected_country,
    NCD_region = selected_country_region,
    World_bank = selected_country_category,
    iso3 = selected_country_code
  )

##add option to rank interventions by percent change in 40q30 per million dollars spent
#NEW#
country.all3<-left_join(country.all3, changeq30%>%filter(Country==selected_country)%>%select(c(-Country)), by="Code")%>%
  mutate(ICER2=(q30diff/sum_increment_all)*1000000, ICER_rank2=rank(-ICER2))


file.name <- paste0("../new_demography/output/costs/",selected_country,"_ICER_2030",".csv")
#write.csv(country.all3, file.name)

final_all_ICER <- rbind(final_all_ICER, country.all3)

}


```


# All data saved
```{r}
file.name <- paste0("../new_demography/output/costs/All_total_cost_2030_intervention_0830",".csv")
write.csv(final_all_intervention, file.name)

file.name <- paste0("../new_demography/output/costs/All_total_cost_2030_year_0830",".csv")
write.csv(final_all, file.name)

file.name <- paste0("../new_demography/output/All_ICER_2030_0830",".csv") #FOR WILL's CODE
write.csv(final_all_ICER, file.name)

file.name <- paste0("../new_demography/output/costs/All_increment_2030_0830",".csv")
write.csv(increment_all_year, file.name)

```

# clean up: intervention by year: pop
```{r}
head(increment_all_year)
head(final_all)

# mean unit costs by intervention and class
df.uc <- increment_all_year %>% 
  group_by(Country, unique_id, Class) %>% 
  summarize(mean_uc = mean(adjusted_uc)) 

df.uc.class <- df.uc%>% 
  group_by(unique_id, Class) %>% 
  summarize(mean_uc_final = mean(mean_uc))

df.uc.class_wide <- spread(data = df.uc.class, key = Class, value = mean_uc_final)

write.csv(df.uc.class_wide, "../new_demography/output/costs/Mean_UC_0830.csv")

# Population by years and senario
base_pin_df <- increment_all_year %>% 
  group_by(unique_id, year_id) %>% 
  summarize(sum_base_pin = sum(base_pin))

base_pin_df_wide <- spread(data = base_pin_df, key = year_id, value = sum_base_pin)
write.csv(base_pin_df_wide, "../new_demography/output/costs/Baseline_pop_0830.csv")

# adjusted
adjust_pin_df <- increment_all_year %>% 
  group_by(unique_id, year_id) %>% 
  summarize(sum_base_pin = sum(adjust_pin))

adjust_pin_df_wide <- spread(data = adjust_pin_df, key = year_id, value = sum_base_pin)
write.csv(adjust_pin_df_wide, "../new_demography/output/costs/Adjusted_pop_0830.csv")

# Incremental
increment_df <- increment_all_year %>% 
  group_by(unique_id, year_id) %>% 
  summarize(sum_increment = sum(incremental))

increment_df_wide <- spread(data = increment_df, key = year_id, value = sum_increment)
write.csv(increment_df_wide, "../new_demography/output/costs/Incremental_cost_0830.csv")

```


# Check output
```{r}
icer <- read.csv("../new_demography/output/All_ICER_2030_0830.csv" , stringsAsFactors=F)

#icer<-final_all_ICER

#codes <- colsplit(icer$Code, ".", names = c("NCD","sub_id"))
icer$NCD <- icer$Code

#Check
#file.name <- paste0("../new_demography/output/costs/All_ICER_2030_re.csv")
#write.csv(icer, file.name)

icer.sum <- icer %>% 
              group_by(NCD) %>%
              summarize(total_increment = sum(sum_increment_all),
                        total_life_year = sum(sum_ly_all))

file.name <- paste0("../new_demography/output/costs/All_increment_ly_sum_0830",".csv")
write.csv(icer.sum, file.name)


icer.sum <- icer %>% 
              group_by(Code) %>%
              summarize(total_increment = sum(sum_increment_all),
                        total_life_year = sum(sum_ly_all))

file.name <- paste0("../new_demography/output/costs/All_increment_ly_sum_re_0830",".csv")
write.csv(icer.sum, file.name)

```


# ICERS by region
```{r ICER}
icer <- read.csv("../new_demography/output/All_ICER_2030_0830.csv", stringsAsFactors=F)
icer$NCD <- icer$Code
self_harm <- c("C2.9", "C2.10", "C2.11", "C2.12")

#weighted by populaiton
load("../new_inputs/PreppedData0819.Rda")
pop<-wpp.in%>%filter(year==2019)%>%group_by(location_name)%>%
  summarise(pop=sum(Nx))%>%rename(Country=location_name)

icer<-left_join(icer, pop, by="Country")
# By region
icer.int <- icer %>%
                group_by(Code, NCD_region) %>%
                summarize(
                  total_ly_gained = sum(sum_ly_all),
                  total_increment = sum(sum_increment_all),
                  mean_q30 = weighted.mean(q30diff, pop)) 

icer.int$icer <- icer.int$total_increment / icer.int$total_ly_gained
icer.int$icer2 <- (icer.int$mean_q30/icer.int$total_increment)*1e6

library(reshape2)
ly_wide <- dcast(data = icer.int, formula = Code ~ NCD_region,
                 value.var = "icer") %>% 
  filter(Code %!in%  self_harm) %>% 
  select(Code, 
         "Latin America and Caribbean",
         "Central and Eastern Europe",
         "Central Asia, Middle East and North Africa",
         "Sub-Saharan Africa",
         "South Asia",
         "East and South East Asia",
         "Oceania"
         )

ly_wide2 <- dcast(data = icer.int, formula = Code ~ NCD_region,
                 value.var = "icer2") %>% 
  filter(Code %!in%  self_harm) %>% 
  select(Code, 
         "Latin America and Caribbean",
         "Central and Eastern Europe",
         "Central Asia, Middle East and North Africa",
         "Sub-Saharan Africa",
         "South Asia",
         "East and South East Asia",
         "Oceania"
         )

write.csv(ly_wide, "../new_demography/output/costs/ICER_region_NCD_0830.csv") #ICER by DALYs
write.csv(ly_wide2, "../new_demography/output/costs/ICER_region_NCD2_0830.csv") #ICER by 40q30


```

```{r ICER ranking}

icer.int2 <- icer.int %>% group_by(NCD_region) %>% 
  filter(Code %!in%  self_harm) %>% 
  mutate(rank = rank(icer), rank2=rank(-icer2))

rank_wide <- dcast(data = icer.int2, formula = Code ~ NCD_region,
                 value.var = "rank") %>% 
  select(Code, 
         "Latin America and Caribbean",
         "Central and Eastern Europe",
         "Central Asia, Middle East and North Africa",
         "Sub-Saharan Africa",
         "South Asia",
         "East and South East Asia",
         "Oceania"
         )

rank_wide2 <- dcast(data = icer.int2, formula = Code ~ NCD_region,
                 value.var = "rank2") %>% 
  select(Code, 
         "Latin America and Caribbean",
         "Central and Eastern Europe",
         "Central Asia, Middle East and North Africa",
         "Sub-Saharan Africa",
         "South Asia",
         "East and South East Asia",
         "Oceania"
         )

write.csv(rank_wide, "../new_demography/output/costs/Rank_region_NCD_0830.csv") #Ranked by DALYs
write.csv(rank_wide2, "../new_demography/output/costs/Rank_region_NCD2_0830.csv") #Ranked by 40q30


```


```{r ICER percent}
gdp <- read.csv("../new_inputs/GDP_per_2017_NCD_region_0608.csv")


gdp <- gdp %>% 
  select(NCD_region, GDP_2017_capita)

icer_long_gdp <- merge(gdp, icer.int,  by.x = "NCD_region", by.y = "NCD_region", all.y = T)

# Variable generate
icer_long_gdp$ICER_percent <-  icer_long_gdp$icer / icer_long_gdp$GDP_2017_capita

#Wide
percent_wide <- dcast(data = icer_long_gdp, formula = Code ~ NCD_region,
                 value.var = "ICER_percent")

# Re-arrange
percent_wide2 <- percent_wide %>% 
  filter(Code %!in%  self_harm) %>% 
  select(Code, 
         "Latin America and Caribbean",
         "Central and Eastern Europe",
         "Central Asia, Middle East and North Africa",
         "Sub-Saharan Africa",
         "South Asia",
         "East and South East Asia",
         "Oceania"
         )
# Save
write.csv(percent_wide2, "../new_demography/output/costs/Percent_region_NCD_0830.csv")

```

#Figure 2 (for paper)

``` {r Heat Map}

library(dplyr)
library(tidyr)
library(ggplot2)
library(heatmaply)
library(RColorBrewer)

setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

#rank by 40q30
#rank<-rank_wide2
rank<-read.csv("../new_demography/output/costs/Rank_region_NCD2_0830.csv")%>%select(-c(X)) 
#rank_daly<-rank_wide
rank_daly<-read.csv("../new_demography/output/costs/Rank_region_NCD_0830.csv")%>%select(-c(X)) 
#color code by DALYs
#icer<-percent_wide2
icer<-read.csv("../new_demography/output/costs/Percent_region_NCD_0830.csv")%>%select(-c(X)) 
#compare the two icer metrics for double legend
#icer2<-ly_wide2
icer2<-read.csv("../new_demography/output/costs/ICER_region_NCD2_0830.csv")%>%select(-c(X)) 

dt<-rank%>%gather(NCD_region, rank, -Code)
dt<-left_join(dt, rank_daly%>%gather(NCD_region, rank, -Code)%>%rename(rank_daly=rank), by=c("Code", "NCD_region"))
dt<-left_join(dt, icer%>%gather(NCD_region, icer_gdp, -Code), by=c("Code", "NCD_region"))
dt<-left_join(dt, icer2%>%gather(NCD_region, icer_40q30, -Code), by=c("Code", "NCD_region"))

#exclude intersectoral
dt<-dt%>%filter(Code<5)

#manually ordering interventions
dt$factor[dt$Code==1.2]<-16
dt$factor[dt$Code==2.4]<-15
dt$factor[dt$Code==2.6]<-14
dt$factor[dt$Code==2.5]<-13
dt$factor[dt$Code==2.7]<-12
dt$factor[dt$Code==2.14]<-11
dt$factor[dt$Code==2.3]<-10
dt$factor[dt$Code==3.1]<-9
dt$factor[dt$Code==3.2]<-8
dt$factor[dt$Code==3.3]<-7
dt$factor[dt$Code==3.4]<-6
dt$factor[dt$Code==3.5]<-5
dt$factor[dt$Code==4.1]<-4
dt$factor[dt$Code==4.2]<-3
dt$factor[dt$Code==4.3]<-2
dt$factor[dt$Code==4.4]<-1

dt$NCD_region<-gsub("\\.", " ", dt$NCD_region)
dt$NCD_region[dt$NCD_region=="Sub Saharan Africa"]<-"Sub-Saharan Africa"
dt$NCD_region[dt$NCD_region=="Central Asia  Middle East and North Africa"]<-"Central Asia, Middle East and North Africa"

dt$factor2[dt$NCD_region=="Latin America and Caribbean"]<-1
dt$factor2[dt$NCD_region=="Central and Eastern Europe"]<-2
dt$factor2[dt$NCD_region=="Central Asia, Middle East and North Africa"]<-3
dt$factor2[dt$NCD_region=="Sub-Saharan Africa"]<-4
dt$factor2[dt$NCD_region=="South Asia"]<-5
dt$factor2[dt$NCD_region=="East and South East Asia"]<-6
dt$factor2[dt$NCD_region=="Oceania"]<-7

names <- read.csv("../new_inputs/names.csv", header=FALSE, stringsAsFactors=FALSE)
names$V1<-as.numeric(gsub("C", "", names$V1))
names<-names%>%rename(Code=V1, Interv=V2)

allints<-left_join(dt, names, by="Code")  

##compare ICER as a share of GDP with 40q30 icer
ggplot(dt, aes(y=log(icer_40q30), x=log(icer_gdp)))+
  geom_point()+
  xlab("Log(ICER as a share of GDP per capita)")+
  ylab("Log(% change in 40q30 \nper $1 million)")+
  geom_smooth(method='lm', formula= y~x, se=F)

#ggsave("Figures/AdditionalFigure_option_0819.png", height = 7, width = 9, units = "in")

ggplot(dt, aes(y=log(icer_40q30), x=log(icer_gdp), color=NCD_region))+
  geom_point()+
  xlab("Log(ICER as a share of GDP per capita)")+
  ylab("Log(% change in 40q30 \nper $1 million)")+
  geom_smooth(method='lm', formula= y~x, se=F)

#ggsave("Figures/AdditionalFigure_option2_0819.png", height = 7, width = 9, units = "in")


dt$text<-paste0("(",dt$rank_daly,",",dt$rank,")")

library(ggrepel)

ggplot(dt%>%group_by(Code)%>%summarise(icer_40q30=mean(icer_40q30), icer_gdp = mean(icer_gdp),
                                       rank=round(mean(rank)), rank_daly=round(mean(rank_daly)))%>%
         mutate(text=paste0("(",rank_daly,",",rank,")")), aes(y=log(icer_40q30), x=log(icer_gdp)))+
  geom_point()+
  xlab("Log(ICER as a share of GDP per capita)")+
  ylab("Log(% change in 40q30 \nper $1 million)")+
  geom_label_repel(aes(label=text),box.padding = 0.1, point.padding=0.1, max.overlaps = 15)

ggsave("Figures/Appendix-figureA2_0830.png", height = 7, width = 11, units = "in")


fit<-lm(log(icer_40q30)~log(icer_gdp), data=dt)
fit
fit$coefficients[1]
fit$coefficients[2]

exp(fit$coefficients[1]+fit$coefficients[2]*log(0.25))
exp(fit$coefficients[1]+fit$coefficients[2]*log(0.5))
exp(fit$coefficients[1]+fit$coefficients[2]*log(0.99))
exp(fit$coefficients[1]+fit$coefficients[2]*log(2.3))


##

allints$colorbreak<-NA
allints$colorbreak[allints$icer_gdp<0.25]<-"<0.25 (>0.003%)"
allints$colorbreak[allints$icer_gdp>=0.25 & allints$icer_gdp<0.5]<-"0.25-0.49 (0.003-0.001%)"
allints$colorbreak[allints$icer_gdp>=0.5 & allints$icer_gdp<1]<-"0.5-0.99 (0.001-0.0007%)"
allints$colorbreak[allints$icer_gdp>=1 & allints$icer_gdp<=2.3]<-"1-2.3 (0.0007-0.0003%)"
allints$colorbreak[allints$icer_gdp>2.3]<-">2.3 (<0.0003%)"

#my_breaks<-c(0.1,0.3,0.5,1,1.5,2)

  ggplot(allints, aes(x=reorder(NCD_region,factor2), y=reorder(Interv,factor), fill=colorbreak))+
  geom_tile()+geom_text(aes(label=rank), color="black")+
  theme(axis.text.x=element_text(size=12, angle=45, hjust=0), axis.text.y=element_text(size=12))+ 
  #scale_fill_gradientn(colors=c("#5d9976","#b9d780", "#feea83","#faa175", "#f8696b", "#800000"), values=c(0,0.5,1))+ #labels=c(0.25,0.5,1.0,1.5,2.0,2.5))+
  scale_fill_manual(values=c("<0.25 (>0.003%)"="#5d9976",
                             "0.25-0.49 (0.003-0.001%)"="#b9d780",
                             "0.5-0.99 (0.001-0.0007%)"="#feea83",
                            "1-2.3 (0.0007-0.0003%)"="#faa175",
                            ">2.3 (<0.0003%)"="#f8696b"
                             ),
                    limits=c("<0.25 (>0.003%)",
                             "0.25-0.49 (0.003-0.001%)","0.5-0.99 (0.001-0.0007%)",
                             "1-2.3 (0.0007-0.0003%)",">2.3 (<0.0003%)"))+
  scale_x_discrete(position="top")+
  labs(fill="ICER as a share of \nGDP per capita \n(% change in 40q30 \nper $1 million)")+
  #theme(legend.position = "bottom",
  #      legend.direction="horizontal")+
  theme(axis.title.x=element_blank(),axis.title.y=element_blank() )
  
  ggsave("Figures/Figure2_0830.png", height = 7, width = 9, units = "in")

```

```{r Table 4 for paper}
#Typical annual case volume per 100,000 population
load("../new_inputs/PreppedData0819.Rda")
total_pop<-wpp.in%>%filter(year==2019)%>%group_by(location_name)%>%
  summarise(pop=sum(Nx))%>%rename(Country=location_name)

# population in need (Output from Will)
load("../new_demography/output/icer_rank_output_0830.Rda")
all.pin$unique<-paste0(all.pin$Code,"_",all.pin$sub_id)

codelist<-read.csv("../new_inputs/caseload.csv", stringsAsFactors = F)
codelist$unique<-paste0(codelist$NCD,"_",codelist$sub_id)

cases<-all.pin%>%filter(group=="Baseline", unique%in%codelist$unique, year_id==2020)%>%group_by(Code, location_name)%>%summarise(cases=sum(pin/Coverage))

cases<-left_join(cases, total_pop%>%rename(location_name=Country), by="location_name")

any(is.na(cases))

caseload<-cases%>%group_by(Code)%>%summarise(volume_per1e5pop=(sum(cases)/sum(pop))*100000)
intnames<-read.csv("../new_inputs/names.csv", stringsAsFactors = F)
intnames$Code<-as.numeric(sub('C','',intnames$Code))

table4<-left_join(caseload, intnames, by="Code")

write.csv(table4, "Figures/Table4_0830.csv")


#alt#
cases2<-all.pin%>%filter(group=="Baseline", year_id==2020)%>%group_by(Code, location_name)%>%summarise(cases=sum(pin/Coverage))
cases2<-left_join(cases2, total_pop%>%rename(location_name=Country), by="location_name")

any(is.na(cases2))

caseload2<-cases2%>%group_by(Code)%>%summarise(volume_per1e5pop=(sum(cases)/sum(pop))*100000)
intnames<-read.csv("../new_inputs/names.csv", stringsAsFactors = F)
intnames$Code<-as.numeric(sub('C','',intnames$Code))

table4.2<-left_join(caseload2, intnames, by="Code")

#write.csv(table4.2, "Figures/Table4.2_0819.csv")


```

```{r Talbe A2 for appendix}

#Cost-effectiveness of NCD4 and intersectoral

#add intersectoral interventions for deaths, dalys, 40q30 - not costs

load("../new_demography/output/icer_rank_output_0830.Rda")
icer <- read.csv("../new_demography/output/All_ICER_2030_0830.csv", stringsAsFactors=F)
icer<-left_join(icer, total_pop, by="Country")

tableA2<-icer%>%group_by(Code)%>%summarise(increment.billions=sum(sum_increment_all)/1e9, 
                                           DALYs.ave.millions=sum(sum_ly_all)/1e6,
                                           Cost.per.DALY.ave=sum(sum_increment_all)/sum(sum_ly_all),
                                           Percent.change.40q30 = weighted.mean(q30diff,pop),
                                           Cost.per.diff40q30 = sum(sum_increment_all)/weighted.mean(q30diff,pop))

names <- read.csv("../new_inputs/names.csv", header=TRUE, stringsAsFactors=FALSE)
names$Code<-as.numeric(gsub("C", "", names$Code))

tableA2<-left_join(tableA2, names, by="Code")

tableA2<-tableA2[order(tableA2$Cost.per.diff40q30, decreasing=FALSE),]

##add deaths averted
deaths<-dadt.all%>%group_by(Code)%>%summarise(deaths.thousands = sum(Deaths.Avert)/1000)
tableA2<-left_join(deaths, tableA2, by="Code")
tableA2$Cost.per.death<-(tableA2$increment.billions*1e9)/(tableA2$deaths.thousands*1000)


write.csv(tableA2, "../new_demography/Figures/Appendix-tableA3_0830.csv")

```
